Update upstream Refactor and changelog
