+++ pyproject.toml
@@ -1,7 +1,7 @@
 [project]
 name = "comfyui-image-saver"
 description = "Save images with generation metadata compatible with Civitai. Works with png, jpeg and webp. Stores LoRAs, models and embeddings hashes for resource recognition."
-version = "1.5.1"
+version = "1.5.2"
 license = { file = "LICENSE" }
 dependencies = ["piexif"]
